# Conflicts:
#	GPFramework/game/data/testmapdata.ini
#	GPFramework/game/imgui.ini
#	GPFramework/source/GP Framework/GP Framework.vcxproj
#	GPFramework/source/GP Framework/GP Framework.vcxproj.filters
#	GPFramework/source/GP Framework/animatedentity.cpp
#	GPFramework/source/GP Framework/animatedentity.h
#	GPFramework/source/GP Framework/elementtype.h
#	GPFramework/source/GP Framework/entity.cpp
#	GPFramework/source/GP Framework/entity.h
#	GPFramework/source/GP Framework/fletchersplayer.cpp
#	GPFramework/source/GP Framework/fletchersplayer.h
#	GPFramework/source/GP Framework/fletchersscene.cpp
#	GPFramework/source/GP Framework/fletchstestblock.cpp
#	GPFramework/source/GP Framework/fletchstestblock.h
#	GPFramework/source/GP Framework/game.cpp
#	GPFramework/source/GP Framework/hud.cpp
#	GPFramework/source/GP Framework/hud.h
#	GPFramework/source/GP Framework/level.cpp
#	GPFramework/source/GP Framework/level.h
#	GPFramework/source/GP Framework/levelmanager.cpp
#	GPFramework/source/GP Framework/levelmanager.h
#	GPFramework/source/GP Framework/levelparser.cpp
#	GPFramework/source/GP Framework/levelparser.h
#	GPFramework/source/GP Framework/scenebbmainmenu.cpp
#	GPFramework/source/GP Framework/scenebbmainmenu.h

+++ GPFramework/source/GP Framework/sprite.h
@@ -19,10 +19,10 @@ public:
 	int GetWidth() const;
 	int GetHeight() const;
 
-	void SetX(int x);
-	int GetX() const;
-	void SetY(int y);
-	int GetY() const;
+	void SetX(float x);
+	float GetX() const;
+	void SetY(float y);
+	float GetY() const;
 
 	void SetAngle(float angle);
 	float GetAngle() const;
@@ -52,8 +52,8 @@ public:
 
 protected:
 	Texture* m_pTexture;
-	int m_x;
-	int m_y;
+	float m_x;
+	float m_y;
 
 	float m_angle;
 	int m_centerX;
